Create a separate subfolder for DB helper classes and functions
