+++ src/track/trackid.h
@@ -2,7 +2,7 @@
 #define TRACKID_H
 
 
-#include "util/dbid.h"
+#include "util/db/dbid.h"
 
 
 class TrackId: public DbId {
